# Conflicts:
#	Assets/Src/FirstLight/Game/Presenters/DeathmatchDeadScreenPresenter.cs
#	Assets/Src/FirstLight/Game/Presenters/GameCountdownScreenPresenter.cs
#	Assets/Src/FirstLight/Game/Presenters/HomeScreenPresenter.cs
#	Assets/Src/FirstLight/Game/Presenters/MatchmakingLoadingScreenPresenter.cs
#	Assets/Src/FirstLight/Game/Presenters/RoomJoinCreateScreenPresenter.cs
#	Assets/Src/FirstLight/Game/Views/MainMenuViews/PlayerNameInputFieldView.cs
#	Assets/Src/FirstLight/Game/Views/MainMenuViews/StandingsHolderView.cs

+++ Quantum/quantum_code/quantum.code/Components/SpecialShieldedCharge.cs
@@ -77,7 +77,7 @@ namespace Quantum
 				ChargeStartPos = attackerPosition,
 				ChargeEndPos = targetPosition,
 				ChargeStartTime = f.Time,
-				PowerAmount = special.PowerAmount
+				PowerAmount = (uint)special.PowerAmount
 			};
 			
 			QuantumHelpers.LookAt2d(f, e, targetPosition);
